requirements.txt
uuid
urllib3
requests >= 2.4.2
rich
python-gitlab
